+++ CHANGELOG.md
@@ -80,3 +80,6 @@
 -   An important bug fix in README.md
 -   Made [fortnite_api_io](https://github.com/vanxh/fortnite_api_io) library a seperate package.
 -   Survivors squads getter in campaign profile.
+-   FORT Stats getter in campaign profile.
+-   Curve table reader.
+-   Power level calculator in campaign profile.